refactor: 맵 요소를 선택하는 커스텀 훅 로직을 단일 선택과 드래그 선택로 분리
